all lelievre ok

+++ libft/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: pwaters <pwaters@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/03 11:11:16 by pwaters           #+#    #+#             */
-/*   Updated: 2021/12/02 11:31:48 by pwaters          ###   ########.fr       */
+/*   Updated: 2021/12/09 12:07:24 by pwaters          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,13 @@
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	int				i;
+	size_t			i;
 	unsigned char	uc;
 	char			*ptr;
 
 	i = 0;
 	uc = (unsigned char)c;
 	ptr = (char *)b;
-	
 	while (i < len)
 	{
 		*(ptr + i) = uc;
